+++ src/noisepy/seis/io/channelcatalog.py
@@ -304,3 +304,10 @@ def cc_parameters(cc_para, coor, tcorr, ncorr, comp):
 
 
 # TODO: A channel catalog that uses the files in the SCEDC S3 bucket: s3://scedc-pds/FDSNstationXML/
+    
+class MockCatalog(ChannelCatalog):
+    def get_full_channel(self, timespan: DateTimeRange, channel: Channel) -> Channel:
+        return channel
+
+    def get_inventory(self, timespan: DateTimeRange, station: Station) -> obspy.Inventory:
+        return obspy.Inventory()
