+++ src/noisepy/seis/io/utils.py
@@ -14,6 +14,9 @@ from tqdm.autonotebook import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
 
 from noisepy.seis.io.constants import AWS_EXECUTION_ENV
+from datetime import datetime, timezone
+
+from datetimerange import DateTimeRange
 
 S3_SCHEME = "s3"
 FIND_RETRIES = 5
@@ -170,3 +173,9 @@ def remove_nan_rows(a: np.ndarray) -> np.ndarray:
     Remove rows from a 2D array that contain NaN values
     """
     return a[~np.isnan(a).any(axis=1)]
+
+def date_range(month: int, start_day: int, end_day: int, start_hr: int = 0, end_hr: int = 0):
+    return DateTimeRange(
+        datetime(2021, month, start_day, start_hr).replace(tzinfo=timezone.utc),
+        datetime(2021, month, end_day, end_hr).replace(tzinfo=timezone.utc),
+    )
