+++ tests/test_hierarchicalstores.py
@@ -6,7 +6,7 @@ from unittest import mock
 import pytest
 from datetimerange import DateTimeRange
 from fsspec.implementations.local import LocalFileSystem  # noqa F401
-from utils import date_range
+from noisepy.seis.io.utils import date_range
 
 from noisepy.seis.io.hierarchicalstores import PairDirectoryCache
 from noisepy.seis.io.numpystore import NumpyArrayStore, NumpyCCStore
