+++ tests/test_scedc_s3store.py
@@ -3,14 +3,11 @@ from datetime import datetime, timezone
 
 import pytest
 from datetimerange import DateTimeRange
-from test_channelcatalog import MockCatalog
+from noisepy.seis.io.channelcatalog import MockCatalog
 
 from noisepy.seis.io.datatypes import Channel, ChannelType, Station
-from noisepy.seis.io.scedc_s3store import SCEDCS3DataStore, channel_filter
+from noisepy.seis.io.scedc_s3store import SCEDCS3DataStore, channel_filter,timespan1
 
-timespan1 = DateTimeRange(
-    datetime(2022, 1, 2, tzinfo=timezone.utc), datetime(2022, 1, 3, tzinfo=timezone.utc)
-)
 timespan2 = DateTimeRange(
     datetime(2021, 2, 3, tzinfo=timezone.utc), datetime(2021, 2, 4, tzinfo=timezone.utc)
 )
